Small fixes to stop compiler warnings.
git-svn-id: https://svn.code.sf.net/p/kaldi/code/trunk@122 5e6a8d80-dfce-4ca6-a32a-6e07a63d50c8

+++ src/base/kaldi-math-test.cc
@@ -80,10 +80,10 @@ void UnitTestRand() {
       assert(RandPoisson(0.0) == 0);
       std::cout << "Test RandPoisson\n";
       float lambda = RandUniform() * 3.0;  // between 0 and 3.
-      int sum = RandPoisson(lambda) - lambda;  // expected value is zero.
+      double sum = RandPoisson(lambda) - lambda;  // expected value is zero.
       for (int j = 0; ; j++) {
         sum += RandPoisson(lambda) - lambda;
-        if (std::abs((double)sum) < 0.5*sqrt((double)j)) break;
+        if (std::abs(sum) < 0.5*sqrt((double)j)) break;
       }
     }
